move average time inspector to own file
